+++ src/App.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Input from './components/Input';
 
 function App() {
   return (
