+++ src/components/Image/Image.stories.tsx
