+++ src/components/Image/ImageStyle.ts
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
 import { css } from '@emotion/react'
 
-import { ImageProps } from '../types/PropsType'
+import { ImageProps } from '../../types/PropsType'
 
 export default styled.img<ImageProps>`
     width: 100px;
