+++ src/components/Input/Input.tsx
@@ -1,5 +1,5 @@
-import { InputProps } from "../types/PropsType";
-import InputStyle from "../styles/InputStyle";
+import { InputProps } from "../../types/PropsType";
+import InputStyle from "./InputStyle";
 
 export default ({disabled, invalid, sizeI, inputProps}: InputProps) => {
     return(
