+++ src/index.tsx
@@ -4,7 +4,7 @@ import App from './App';
 import { Global } from '@emotion/react';
 import reportWebVitals from './reportWebVitals';
 
-import reset from './styles/reset';
+import globalStyle from './globalStyle';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -12,12 +12,9 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <Global styles={reset}/>
+    <Global styles={globalStyle}/>
     <App />
   </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
